requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.1
openai==1.3.7
chromadb==0.4.18
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==1.1.0
markdown==3.5.1
unstructured==0.11.0
faiss-cpu==1.7.4
requests==2.31.0
beautifulsoup4==4.12.2
google-search-results==2.4.2 